chore: migrate playgroud to features floder
