+++ apps/mobile/app/playground/components/theme-token.tsx
@@ -1,0 +1,5 @@
+import ThemeTokensPlayground from '@/features/playground/components/theme-token';
+
+export default function ThemeTokensPlaygroundPage() {
+  return <ThemeTokensPlayground />;
+}
