+++ apps/mobile/src/features/playground/components/style.ts
@@ -1,0 +1,8 @@
+import { createStyles } from '@/theme';
+
+export const useStyles = createStyles((token) => ({
+  safeAreaView: {
+    backgroundColor: token.colorBgLayout,
+    flex: 1,
+  },
+}));
